# authz: Use conf.Watch in SubRepoPermsClient