@ryanslade 2022-09-06
PTAL, we now reset the cache size correctly